date query is added
What changes you have made?
added feature of query by date
added predicate for dateMatch and a general query func to handle all
query funcs
.....

Why this change is necessary?
As of now, it's a part of the requirement so now users are able to make
a query by date, empId or with both of them
....

What is the impact of the change?
the general query func now handle all queries so no pressure on the
performActioin func..
date filter in working alone and also with empId filter
.....