Query by Beverage feature is  added
What changes you have made?
Now one can make query by beverage, so new bvrgQuery() func is added
that is handled by performQuery() func

README is updated according to the change
.....

Why this change is necessary?
To meet current requirement
....

What is the impact of the change?
It enhances the utility of the tool
.....